requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
pandas==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.335
langchain-community==0.0.6
ollama==0.1.7
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0